251116 patch

+++ Albedo/Utils/NumberUtil.cs
@@ -23,12 +23,12 @@ namespace Albedo.Utils
             return value.ToString("#,0.############################");
         }
 
-        public static double Max(double value1, double value2, double value3)
+        public static decimal Max(decimal value1, decimal value2, decimal value3)
         {
             return Math.Max(value1, Math.Max(value2, value3));
         }
 
-        public static double Min(double value1, double value2, double value3)
+        public static decimal Min(decimal value1, decimal value2, decimal value3)
         {
             return Math.Min(value1, Math.Min(value2, value3));
         }
